+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 - Сhg #37: Remove parameters in `ApplicationRunner` methods `runBootstrap()` and `checkEvents()`, instead are used 
   internal container and config instances (@vjik) 
-- New #36: Add ability to configure container configuration group (@vjik)
-- New #38: Add ability to configure config params group (@vjik)
+- Chg #39: Adapt to Yii configuration groups names convention (@vjik)
+- Chd #39: Add parameters `$useBootstrap`, `$checkEvents` and `$configGroupPostfix` to `ApplicationRunner` constructor
+  and removed methods `withBootstrap()`, `withoutBootstrap()`, `withCheckingEvents()`, `withoutCheckingEvents()` (@vjik) 
 
 ## 1.2.1 November 07, 2022